+++ src/Users/Mypage.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import Nav from "./Nav";
+import Nav from "../Nav";
 import './Mypage.css';
 import axios from "axios";
