+++ src/Users/Signup.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import './Signup.css';
 import axios from 'axios';
-import Nav from './Nav';
+import Nav from '../Nav';
 import { useNavigate } from 'react-router-dom';
 
 const Signup = () => {
